<==== Mercurial data ====>
Branch: nat_traversal

+++ appserver2/src/llutil/util.cpp
@@ -6,7 +6,7 @@ void LLUtil::changeGuidByteOrder(std::string& guid)
 
     int replace[] = {6, 7, 4, 5, 2, 3, 0, 1, 8, 11, 12, 9, 10, 13, 16, 17, 14, 15};
     
-    for (int i = 0; i < sizeof(replace) / sizeof(replace[0]); i++) {
+    for (size_t i = 0; i < sizeof(replace) / sizeof(replace[0]); i++) {
         guid_fixed[i] = guid[replace[i]];
     }
